First pass at parameterising the Vagrant scripts
Allows you to use a Vagrantfile.local to override the default settings.

+++ vagrant/solarnet-dev/solardev.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
+SETUP_FLUXBOX=$1
+
 # Setup .xinitrc to launch Fluxbox
-if [ ! -e ~/.xinitrc ]; then
+if $SETUP_FLUXBOX && [ ! -e ~/.xinitrc ]; then
 	echo 'Configuring Fluxbox in .xinitrc...'
 	echo "exec startfluxbox" > ~/.xinitrc
 fi
